+++ src/main/java/com/yolo/game/event/BetEvent.java
@@ -4,12 +4,10 @@ import com.yolo.game.engine.Player;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @RequiredArgsConstructor
 public class BetEvent implements PlayerEvent {
     private final Player player;
-    private final Integer number ;
-    private final BigDecimal stake;
+    private final int number;
+    private final double stake;
 }
